Implement Categories GET endpoint with MySQL DAO Implement product-by-category endpoint Added GET /categories/{id} product endpoint Phase 1 completed: categories,products, auto, security Phase 2 -Added ShoppingCartController endpoints and DAO stubs start phase 3 cart updates setup fixed put cart endpoint to use request body
Implement shopping cart add product flow, DAO integration, and JWT-secured endpoints
Added product to cart endpoint with auto and response
Updated profile-service.js, for user profile

Phase 3:
Added POST/orders endpoint secured with JWT
Implemented OrderDao and MysqlOrderDao createOrder Logic
Inserted orders and order_line_items using INSERT SELECT
Linked orders to authenticated user via Principal
Cleared shopping cart after successful checkout
Fixed category delete and return 404 when category not found
added update and delete methode
JWT login returns token
Added get profile endpoint for authenticated user
added update profile endpoint
implement profile DAO and MySQL DAO
Implement login, profile, and order creation endpoints
fixed order creation by loading profile via profileDao
added README

updated Category and created orderLineItem class
Implemented the line total display

edited README

+++ backend-api/src/main/java/org/yearup/controllers/CategoriesController.java
@@ -1,3 +1,101 @@
+//package org.yearup.controllers;
+//
+//import org.springframework.beans.factory.annotation.Autowired;
+//import org.springframework.http.HttpStatus;
+//import org.springframework.http.ResponseEntity;
+//import org.springframework.security.access.prepost.PreAuthorize;
+//import org.springframework.web.bind.annotation.*;
+//import org.springframework.web.server.ResponseStatusException;
+//import org.yearup.data.CategoryDao;
+//import org.yearup.data.ProductDao;
+//import org.yearup.models.Category;
+//import org.yearup.models.Product;
+//
+//import java.util.List;
+//
+//// add the annotations to make this a REST controller
+//// add the annotation to make this controller the endpoint for the following url
+//// http://localhost:8080/categories
+//// add annotation to allow cross site origin requests
+//
+//@RestController
+//@RequestMapping("/categories")
+//@CrossOrigin
+//public class CategoriesController {
+//    private CategoryDao categoryDao;
+//    private ProductDao productDao;
+//
+//
+//    // create an Autowired controller to inject the categoryDao and ProductDao
+//    @Autowired
+//    public CategoriesController(CategoryDao categoryDao, ProductDao productDao) {
+//        this.categoryDao = categoryDao;
+//        this.productDao = productDao;
+//    }
+//
+//    // add the appropriate annotation for a get action
+//    @GetMapping
+//    public List<Category> getAll() {
+//        // find and return all categories
+//        return categoryDao.getAllCategories();
+//    }
+//
+//    // add the appropriate annotation for a get action
+//    @GetMapping("/{id}")
+//    public ResponseEntity<Category> getById(@PathVariable int id) {
+//        Category category = categoryDao.getById(id);
+//
+//        if (category == null) {
+//            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Category not found");
+//        }
+//        return ResponseEntity.ok(category);
+//
+//    }
+//
+//    @GetMapping("/{id}/products")
+//    public List<Product> getProductsByCategory(@PathVariable int id) {
+//        return productDao.search(id, null, null, null);
+//    }
+//
+//    // the url to return all products in category 1 would look like this
+//    // https://localhost:8080/categories/1/products
+//
+//
+//    // add annotation to call this method for a POST action
+//    // add annotation to ensure that only an ADMIN can call this function
+//    @PostMapping
+//    @PreAuthorize("hasRole('ADMIN')")
+//    public ResponseEntity<Category> addCategory(@RequestBody Category category) {
+//        Category created = categoryDao.create(category);
+//        return ResponseEntity.status(HttpStatus.CREATED).body(created);
+//    }
+//
+//    // add annotation to call this method for a PUT (update) action - the url path must include the categoryId
+//    // add annotation to ensure that only an ADMIN can call this function
+//    @PutMapping("/{id}")
+//    @PreAuthorize("hasRole('ADMIN')")
+//    @ResponseStatus(HttpStatus.NO_CONTENT)
+//    public void updateCategory(@PathVariable int id, @RequestBody Category category) {
+//        categoryDao.update(id, category);
+//        // update the category by id
+//    }
+//
+//
+//    // add annotation to call this method for a DELETE action - the url path must include the categoryId
+//    // add annotation to ensure that only an ADMIN can call this function
+//    @DeleteMapping("/{id}")
+//    @PreAuthorize("hasRole('ADMIN')")
+//    @ResponseStatus(HttpStatus.NO_CONTENT)
+//    public void deleteCategory(@PathVariable int id) {
+//        categoryDao.delete(id);
+//        // delete the category by id
+//    }
+//}
+
+
+
+
+
 package org.yearup.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,55 +115,63 @@ import java.util.List;
 // add the annotation to make this controller the endpoint for the following url
 // http://localhost:8080/categories
 // add annotation to allow cross site origin requests
-
+@CrossOrigin
 @RestController
 @RequestMapping("/categories")
-@CrossOrigin
-public class CategoriesController {
+public class CategoriesController
+{
     private CategoryDao categoryDao;
     private ProductDao productDao;
 
-
-    // create an Autowired controller to inject the categoryDao and ProductDao
+    // create an Autowired controller to inject the categoryDao and ProductDaowi
     @Autowired
-    public CategoriesController(CategoryDao categoryDao, ProductDao productDao) {
+    public CategoriesController(CategoryDao categoryDao, ProductDao productDao)
+    {
         this.categoryDao = categoryDao;
         this.productDao = productDao;
     }
 
     // add the appropriate annotation for a get action
     @GetMapping
-    public List<Category> getAll() {
+    @PreAuthorize("permitAll()")
+    public List<Category> getAll()
+    {
         // find and return all categories
         return categoryDao.getAllCategories();
     }
 
     // add the appropriate annotation for a get action
     @GetMapping("/{id}")
-    public ResponseEntity<Category> getById(@PathVariable int id) {
-        Category category = categoryDao.getById(id);
-
-        if (category == null) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Category not found");
+    @PreAuthorize("permitAll()")
+    public Category getById(@PathVariable int id)
+    {
+        // get the category by id
+        Category result = categoryDao.getById(id);
+        if (result == null)
+        {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
         }
-        return ResponseEntity.ok(category);
-
-    }
-
-    @GetMapping("/{id}/products")
-    public List<Product> getProductsByCategory(@PathVariable int id) {
-        return productDao.search(id, null, null, null);
+        return result;
     }
 
     // the url to return all products in category 1 would look like this
     // https://localhost:8080/categories/1/products
-
+    @GetMapping("/{categoryId}/products")
+    @PreAuthorize("permitAll()")
+    public List<Product> getProductsById(@PathVariable int categoryId)
+    {
+        // get a list of product by categoryId
+        return productDao.listByCategoryId(categoryId);
+    }
 
     // add annotation to call this method for a POST action
     // add annotation to ensure that only an ADMIN can call this function
     @PostMapping
-    @PreAuthorize("hasRole('ADMIN')")
-    public ResponseEntity<Category> addCategory(@RequestBody Category category) {
+    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    @ResponseStatus(HttpStatus.CREATED)
+    public ResponseEntity<Category> addCategory(@RequestBody Category category)
+    {
+        // insert the category
         Category created = categoryDao.create(category);
         return ResponseEntity.status(HttpStatus.CREATED).body(created);
     }
@@ -73,21 +179,35 @@ public class CategoriesController {
     // add annotation to call this method for a PUT (update) action - the url path must include the categoryId
     // add annotation to ensure that only an ADMIN can call this function
     @PutMapping("/{id}")
-    @PreAuthorize("hasRole('ADMIN')")
+    @PreAuthorize("hasRole('ROLE_ADMIN')")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void updateCategory(@PathVariable int id, @RequestBody Category category) {
-        categoryDao.update(id, category);
+    public void updateCategory(@PathVariable int id, @RequestBody Category category)
+    {
         // update the category by id
-    }
+        Category existing = categoryDao.getById(id);
+        if (existing == null)
+        {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
 
+        category.setCategoryId(id);
+        categoryDao.update(id, category);
+    }
 
     // add annotation to call this method for a DELETE action - the url path must include the categoryId
     // add annotation to ensure that only an ADMIN can call this function
     @DeleteMapping("/{id}")
-    @PreAuthorize("hasRole('ADMIN')")
+    @PreAuthorize("hasRole('ROLE_ADMIN')")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteCategory(@PathVariable int id) {
-        categoryDao.delete(id);
+    public void deleteCategory(@PathVariable int id)
+    {
         // delete the category by id
+        Category existing = categoryDao.getById(id);
+        if (existing == null)
+        {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
+
+        categoryDao.delete(id);
     }
 }
